requirements.txt
# FLEX-OP Requirements
# Python 3.10+

# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0  # Database URL parsing

# CORS
django-cors-headers>=4.3.0

# Environment variables
python-dotenv>=1.0.0

# API Documentation
drf-yasg>=1.21.7  # Swagger/OpenAPI

# Image handling
Pillow>=10.1.0

# Password validation
django-password-validators>=1.7.1

# Utilities
python-dateutil>=2.8.2

# Development tools (optional)
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.1.0
black>=23.12.0

# Production server
gunicorn>=21.2.0

# Background tasks (opcional - puede usar cron)
celery>=5.3.4
redis>=5.0.1

# Reporting (para PDFs)
reportlab>=4.0.7
